requirements-nemo.txt
langchain-google-genai>=2.0.0
nemoguardrails>=0.10.0
